+++ crates/zng/src/slider.rs
@@ -1,0 +1,13 @@
+//! Slider widget, styles and properties.
+//!
+//! This widget allows selecting a value or range by dragging a selector knob over a range line.
+//!
+//! ```
+//! todo!()
+//! ```
+//!
+//! # Full API
+//!
+//! See [`zng_wgt_slider`] for the full widget API.
+
+pub use zng_wgt_slider::Slider;