+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Add `zng::slider` with `Slider` widget.
 * Fix headless rendering in Wayland. Property `needs_fallback_chrome` now is `false` for headless windows.
 
 # 0.12.7